requirements.txt
boto3==1.26.118
fastapi==0.95.1
mangum==0.17.0
python-dotenv==1.0.0
uvicorn==0.21.1
httpx==0.24.0
pytest==7.3.1
PyJWT==2.6.0